Update index.css
By default flexbox will try to keep all flex items on a single line, shrinking them
down to almost nothing if it has to. However if we want flex items to take up
their full size and wrap to a next line we need the flex-wrap property. By
default flex-wrap is set to nowrap . By simply setting flex-wrap: wrap on
the flex container we tell it to let its flex items wrap to the next line if there is
not enough room for them. This allows the items to maintain their original size
for longer.